+++ middleware/framework/facade/ApplicationFacadeFactoryBean.js
@@ -1,15 +1,15 @@
-import { GetLatestMoviesRepository } from '../repository/movies/getMovies/GetLatestMoviesRepository'
+import { GetLatestMoviesRepository } from '../repository/movies/getLatest/GetLatestMoviesRepository'
 import { GetTrendingMoviesRepository } from '../repository/trending/getTrending/GetTrendingMoviesRepository'
 import { GetTvShowsVideosRepository } from '../repository/tvShows/GetVideos/GetTvShowsVideosRepository'
-import { GetLatestMoviesService } from '../../application/usecases/movies/getMovies/GetLatestMoviesService'
+import { GetLatestMoviesService } from '../../application/usecases/movies/getLatest/GetLatestMoviesService'
 import { GetTrendingMoviesService } from '../../application/usecases/trending/getTrending/GetTrendingMoviesService'
 import { GetTvShowsVideosService } from '../../application/usecases/tv/GetVideos/GetTvShowsVideosService'
 import { GetTrendingMoviesController } from '../controller/trending/getTrending/GetTrendingMoviesController'
-import { GetLatestMoviesController } from '../controller/movies/getMovies/GetLatestMoviesController'
+import { GetLatestMoviesController } from '../controller/movies/getLatest/GetLatestMoviesController'
 import { GetTvShowsVideosController } from '../controller/tv/GetVideos/GetTvShowsVideosController'
 import { GetMovieVideosRepository } from '../repository/movies/getVideos/GetMovieVideosRepository'
-import { GetMovieVideosService } from '../../application/usecases/movies/getMovieVideos/GetMovieVideosService'
-import { GetMovieVideosController } from '../controller/movies/getMovieVideos/GetMovieVideosController'
+import { GetMovieVideosService } from '../../application/usecases/movies/getVideos/GetMovieVideosService'
+import { GetMovieVideosController } from '../controller/movies/getVideos/GetMovieVideosController'
 import * as Globalconfiguration from '../modules/security/GlobalConfiguration'
 
 const accessToken = Globalconfiguration.theMovieDBConfigEnv
